+++ x/fungible/keeper/deposits_test.go
@@ -6,7 +6,9 @@ import (
 
 	"cosmossdk.io/math"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
+	"github.com/zeta-chain/zetacore/cmd/zetacored/config"
 	"github.com/zeta-chain/zetacore/pkg/chains"
 	"github.com/zeta-chain/zetacore/pkg/coin"
 	"github.com/zeta-chain/zetacore/testutil/contracts"
@@ -351,4 +353,19 @@ func TestKeeper_ZRC20DepositAndCallContract(t *testing.T) {
 		require.NoError(t, err)
 		require.EqualValues(t, int64(0), balance.Int64())
 	})
+}
+
+func TestKeeper_DepositCoinZeta(t *testing.T) {
+	k, ctx, sdkk, _ := testkeeper.FungibleKeeper(t)
+	to := sample.EthAddress()
+	amount := big.NewInt(1)
+	zetaToAddress := sdk.AccAddress(to.Bytes())
+
+	b := sdkk.BankKeeper.GetBalance(ctx, zetaToAddress, config.BaseDenom)
+	require.Equal(t, int64(0), b.Amount.Int64())
+
+	err := k.DepositCoinZeta(ctx, to, amount)
+	require.NoError(t, err)
+	b = sdkk.BankKeeper.GetBalance(ctx, zetaToAddress, config.BaseDenom)
+	require.Equal(t, amount.Int64(), b.Amount.Int64())
 }