+++ changelog.md
@@ -57,6 +57,7 @@
 * [1967](https://github.com/zeta-chain/node/pull/1967) - improve crosschain module coverage
 * [1955](https://github.com/zeta-chain/node/pull/1955) - improve emissions module coverage
 * [1941](https://github.com/zeta-chain/node/pull/1941) - add unit tests for zetabridge package
+* [1985](https://github.com/zeta-chain/node/pull/1985) - improve fungible module coverage
 
 ### Fixes